+++ Test/AccountTest.cs
@@ -29,7 +29,8 @@ namespace Recurly.Test
                 VatNumber = "my-vat-number",
                 TaxExempt = true,
                 EntityUseCode = "I",
-                CcEmails = "cc1@test.com,cc2@test.com"
+                CcEmails = "cc1@test.com,cc2@test.com",
+                Address = new Address()
             };
 
             string address = "123 Faux Street";
@@ -49,6 +50,13 @@ namespace Recurly.Test
             Assert.Equal("I", acct.EntityUseCode);
             Assert.Equal(address, acct.Address.Address1);
             Assert.False(acct.VatLocationValid);
+        }
+
+        [Fact]
+        public void DontSerializeNullAddress()
+        {
+            var account = new Account("testAcct");
+            account.Address.Should().BeNull();
         }
 
         [Fact]